+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 
 ### Enhancements
 - GLS-41 - Added hashed id attribute to users on Activity Stream endpoint
+- GLS-91 - Add management command for data retention notifications
 
 ### Fixed bugs
 - GLS-91 - Data retentions notification bugfix